refactor: simplify its2 revision update checks
