+++ pkg/controller/instanceset2/update_plan.go
@@ -29,16 +29,16 @@ import (
 	intctrlutil "github.com/apecloud/kubeblocks/pkg/controllerutil"
 )
 
-func newUpdatePlan(its workloads.InstanceSet, instances []*workloads.Instance, updated func(*workloads.InstanceSet, *workloads.Instance) bool) updatePlan {
+func newUpdatePlan(its workloads.InstanceSet, instances []*workloads.Instance, desiredInstances map[string]*workloads.Instance) updatePlan {
 	var instanceList []workloads.Instance
 	for _, inst := range instances {
 		instanceList = append(instanceList, *inst)
 	}
 	return &realUpdatePlan{
-		its:               its,
-		instances:         instanceList,
-		dag:               graph.NewDAG(),
-		isInstanceUpdated: updated,
+		its:              its,
+		instances:        instanceList,
+		desiredInstances: desiredInstances,
+		dag:              graph.NewDAG(),
 	}
 }
 
@@ -59,9 +59,9 @@ var (
 type realUpdatePlan struct {
 	its                  workloads.InstanceSet
 	instances            []workloads.Instance
+	desiredInstances     map[string]*workloads.Instance
 	dag                  *graph.DAG
 	instancesToBeUpdated []*workloads.Instance
-	isInstanceUpdated    func(*workloads.InstanceSet, *workloads.Instance) bool
 }
 
 var _ updatePlan = &realUpdatePlan{}
@@ -82,7 +82,7 @@ func (p *realUpdatePlan) planWalkFunc(vertex graph.Vertex) error {
 	}
 
 	// if instance is the latest version, we do nothing
-	if p.isInstanceUpdated(&p.its, inst) {
+	if isInstanceUpdated(&p.its, inst, p.desiredInstances[inst.Name]) {
 		if !intctrlutil.IsInstanceReady(inst) {
 			return ErrWait
 		}
